zfs: verify kernel compatibility against META
Switch to specifying the Kernel compatibility in the same form as in the
ZFS source’s `META` file. This then allows checking it against what’s in
`META` to remove the need to manually check, and so reduce the cognitive
load and change of mistakes when performing updates.

As the “max” version is indeed a compatible version, we need to
increment that for the comparison when using `versionOlder`.

One side-effect here is that it makes it not possible to override the
specified compatibility. We will “fix” this soon.